requirements-auth.txt
# JWT Authentication System Requirements

# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic[email]>=2.0.0

# JWT handling
pyjwt[crypto]>=2.8.0
cryptography>=41.0.0

# Password hashing
bcrypt>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client for testing
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
